Fixed missing wait; test progress tracker

+++ src/test/java/format/ffmpegTests.java
@@ -17,7 +17,8 @@ public class ffmpegTests {
         if(!ffmpeg.isValidInstallation()) // use bin if no local installation
             ffmpeg = new FFMPEG("bin/ffmpeg.exe", "bin/ffprobe.exe");
 
-        Files.copy(new File("src/test/resources/format/video.mp4").toPath(),input.toPath());
+        if(!input.exists())
+            Files.copy(new File("src/test/resources/format/video.mp4").toPath(),input.toPath());
     }
 
     @AfterClass
